@@ -4,7 +4,8 @@ import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+
 import java.io.Serializable;
 import java.sql.Timestamp;
 
@@ -57,7 +58,7 @@ public class Image implements Serializable {
     /**
      * 用户
      */
-    @ManyToOne
+    @ManyToOne(targetEntity = User.class)
     private User user;
 
     /**
