@@ -6,7 +6,7 @@ import com.nekolr.saber.service.IdSeqService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 
 @Service
